refactor sed components
