+++ explore/src/main/scala/explore/targeteditor/spectralDefinition/emissionLineEditor/EmissionLineEditor.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.targeteditor
+package explore.targeteditor.spectralDefinition.emissionLineEditor
 
 import cats.Order.*
 import cats.syntax.all.*
