+++ explore/src/main/scala/explore/targeteditor/spectralDefinition/emissionLineEditor/package.scala
@@ -1,0 +1,14 @@
+// Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+/*
+ * Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
+ * For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+package explore.targeteditor.spectralDefinition.emissionLineEditor
+
+import lucuma.core.util.NewType
+
+object AddDisabled extends NewType[Boolean]
+type AddDisabled = AddDisabled.Type
